Add files via upload
Add weather data csv from 01/01/1980 to 09/20/2025. 

I figured we didn't need anything before that because it might start to get very computationally expensive and for those of us that do not have a GPU, I figured this was enough. But also, if we need, I can grab the whole dataset once we've finished the pipeline. Use this as needed.